meta: replace setup.py by modern pyproject.toml
