--- setup.cfg
@@ -1,25 +1,0 @@
-[metadata]
-name = sopel-slap
-version = 0.1.0
-description = Plugin for Sopel that lets users slap each other in fun ways
-author = dgw
-author_email = dgw@technobabbl.es
-url = https://github.com/sopel-irc/sopel-slap
-license = Eiffel Forum License, version 2
-classifiers =
-    Intended Audience :: Developers
-    Intended Audience :: System Administrators
-    License :: Eiffel Forum License (EFL)
-    License :: OSI Approved :: Eiffel Forum License
-    Topic :: Communications :: Chat :: Internet Relay Chat
-
-[options]
-packages = find:
-zip_safe = false
-include_package_data = true
-install_requires =
-    sopel>=7.0,<8
-
-[options.entry_points]
-sopel.plugins =
-    slap = sopel_slap
